Use UseState and separete in components
We add the use the event of the button to mark the dish as favorte, and
the button to mark a dish as shared.

we use usestate to change the state of the button when it is pressed.

we separate the mainPage in components, and crate a new file to the
button component.

Co-authored-by: javs1 <javierverduzco13@gmail.com>
Co-authored-by: fpreciado3 <fpreciado3@ucol.mx>
Co-authored-by: WEROSKI <oheredia2@ucol.mx>
Co-authored-by: vanessa-covarrubias-nava <vanessa.covarrubias.nava@gmail.com>