Implement Jest tests for Button
Add comprehensive Jest tests to ensure the functionality and reliability of [Button].

We tried to test the functionality of the button by simulating the clic
event, but we couldn´t do it XD. We tried to use the fireEvent funtion
but it didn´t work. we also tried to use the simulation of the clic.